requirements.txt
flask
scikit-learn
numpy
joblib
Flask-Login
Flask-WTF
Flask-SQLAlchemy
email-validator
python-dotenv
pandas